+++ test/integration/build_requires/profile_build_requires_test.py
@@ -3,7 +3,7 @@ import platform
 import textwrap
 import unittest
 
-from conans.paths import CONANFILE
+from conan.internal.paths import CONANFILE
 from conan.test.utils.tools import TestClient, GenConanfile
