+++ test/integration/command/install/test_install_transitive.py
@@ -4,7 +4,7 @@ import pytest
 
 from conans.model.info import load_binary_info
 from conans.model.recipe_ref import RecipeReference
-from conans.paths import CONANFILE_TXT, CONANINFO
+from conan.internal.paths import CONANFILE_TXT, CONANINFO
 from conan.test.utils.tools import TestClient,  GenConanfile
 from conans.util.files import save, load
