+++ test/integration/command/source_test.py
@@ -6,7 +6,7 @@ from collections import OrderedDict
 
 import pytest
 
-from conans.paths import CONANFILE
+from conan.internal.paths import CONANFILE
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient, TestServer
 from conans.util.files import save
