+++ test/integration/command/upload/upload_complete_test.py
@@ -3,7 +3,7 @@ import unittest
 
 from requests import ConnectionError
 
-from conans.paths import CONAN_MANIFEST
+from conan.internal.paths import CONAN_MANIFEST
 from conan.test.utils.tools import (TestClient, TestRequester, TestServer,
                                      GenConanfile)
 from conans.util.env import environment_update
