+++ test/integration/remote/multi_remote_test.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 from time import sleep
 
 from conans.model.recipe_ref import RecipeReference
-from conans.paths import CONANFILE
+from conan.internal.paths import CONANFILE
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient, TestServer
