+++ test/integration/settings/settings_override_test.py
@@ -3,7 +3,7 @@ import os
 import pytest
 
 from conans.model.recipe_ref import RecipeReference
-from conans.paths import CONANINFO
+from conan.internal.paths import CONANINFO
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient
 from conans.util.files import load
