+++ test/integration/tgz_macos_dot_files_test.py
@@ -10,7 +10,7 @@ import pytest
 
 from conans.client.remote_manager import uncompress_file
 from conans.model.recipe_ref import RecipeReference
-from conans.paths import EXPORT_SOURCES_TGZ_NAME
+from conan.internal.paths import EXPORT_SOURCES_TGZ_NAME
 from conan.test.utils.tools import TestClient, NO_SETTINGS_PACKAGE_ID
