+++ test/unittests/paths/user_home_test.py
@@ -2,7 +2,7 @@ import os
 import platform
 from pathlib import Path
 
-from conans.paths import get_conan_user_home, DEFAULT_CONAN_HOME
+from conan.internal.paths import get_conan_user_home, DEFAULT_CONAN_HOME
 from conan.test.utils.test_files import temp_folder
 from conans.util.files import chdir
